create user authentication and separate user data

+++ backend/src/controllers/items.ts
@@ -2,19 +2,50 @@ import { RequestHandler } from "express";
 import mongoose from "mongoose";
 import createHttpError from "http-errors";
 import ItemModel from "../models/item";
-import { createDeflate } from "zlib";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getItems: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const items = await ItemModel.find({ isDone: false }).exec();
+        assertIsDefined(authenticatedUserId);
+        const items = await ItemModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(items);
     } catch (error) {
         next(error);
     }
 };
 
+
+
+export const getItem: RequestHandler = async (req, res, next) => {
+    const itemId = req.params.itemId;
+    const authenticatedUserId = req.session.userId;
+
+    try {
+        assertIsDefined(authenticatedUserId);
+        if (!mongoose.isValidObjectId(itemId)) {
+            throw createHttpError(400, "Invalid Item Id");
+        }
+
+        const item = await ItemModel.findById(itemId).exec();
+
+        if (!item) {
+            throw createHttpError(404, "Item not found");
+        }
+
+        if (!item.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this Item");
+        }
+
+        res.status(200).json(item);
+    } catch (error) {
+        next(error);
+    }
+};
+
+
 interface CreateItemBody {
-    id: number,
     isDone?: boolean,
     content: string,
     createDate: number,
@@ -27,17 +58,21 @@ interface CreateItemBody {
 // 3. req.body => define the type of body using interface that I created above
 // 4. req.query => unkown
 export const createItems: RequestHandler<unknown, unknown, CreateItemBody, unknown> = async (req, res, next) => {
-    const id = req.body.id;
     const isDone = req.body.isDone;
     const content = req.body.content;
     const createDate = req.body.createDate;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        // console.log('Creating new item with data:', { userId: authenticatedUserId, isDone, content, createDate });
+        assertIsDefined(authenticatedUserId);
         if (!content) {
             throw createHttpError(400, "Must include Content");
         }
+     
         const newItem = await ItemModel.create( {
-            id: id,
+            userId: authenticatedUserId,
             isDone: isDone,
             content: content,
             createDate: createDate,
@@ -51,15 +86,23 @@ export const createItems: RequestHandler<unknown, unknown, CreateItemBody, unkno
 
 export const completeItem: RequestHandler = async(req, res, next) => {
     const itemId = req.params.itemId;
+    const authenticatedUserId = req.session.userId;
     console.log(itemId);
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(itemId)) {
             throw createHttpError(400, "Invalid Item Id");
         }
         const item = await ItemModel.findById(itemId).exec();
+
         if (!item) {
             throw createHttpError(400, "Item Not Found");
         }
+
+        if (!item.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this Item");
+        }
+
         const result = await ItemModel.updateOne(
             {_id: itemId},
             {$set : {isDone: true}}
@@ -73,14 +116,22 @@ export const completeItem: RequestHandler = async(req, res, next) => {
 
 export const deleteItem: RequestHandler = async(req, res, next) => {
     const itemId = req.params.itemId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(itemId)) {
             throw createHttpError(400, "Invalid Item Id !");
         }
         const item = await ItemModel.findById(itemId).exec();
+
         if (!item) {
             throw createHttpError(400, "Item not found!"); 
+        }
+
+        if (!item.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this Item");
         }
 
         await item.deleteOne();
